Attempt: Code refactoring
