+++ packages/app-mobile/components/NoteEditor/CodeMirror/markdownCommands.ts
@@ -6,645 +6,25 @@ import { EditorView, Command } from '@codemirror/view';
 
 import { ListType } from '../types';
 import {
-	SelectionRange, EditorSelection, ChangeSpec, Line, TransactionSpec, EditorState,
-	Text as DocumentText,
+	SelectionRange, EditorSelection, ChangeSpec, Line, TransactionSpec,
 } from '@codemirror/state';
 import { getIndentUnit, indentString, syntaxTree } from '@codemirror/language';
-import { SyntaxNodeRef } from '@lezer/common';
-import RegionSpec from './RegionSpec';
+import {
+	RegionSpec, growSelectionToNode, renumberList,
+	toggleInlineFormatGlobally, toggleRegionFormatGlobally, toggleSelectedLinesStartWith,
+	isIndentationEquivalent, stripBlockquote, tabsToSpaces,
+} from './markdownReformatter';
 import { logMessage } from './webviewLogger';
 
-// Length of the symbol that starts a block quote
-const blockQuoteStartLen = '> '.length;
-const blockQuoteRegex = /^>\s/;
+
 const startingSpaceRegex = /^(\s*)/;
-
-// Specifies the update of a single selection region and its contents
-type SelectionUpdate = { range: SelectionRange; changes?: ChangeSpec };
-
-/** @return the text of [line], ignoring starting blockquote characters. */
-const stripBlockquote = (line: Line): string => {
-	const match = line.text.match(blockQuoteRegex);
-
-	if (match) {
-		return line.text.substring(match[0].length);
-	}
-
-	return line.text;
-};
-
-/** @return a version of [text] with all tabs converted to spaces */
-export const tabsToSpaces = (state: EditorState, text: string): string => {
-	const chunks = text.split('\t');
-	const spaceLen = getIndentUnit(state);
-	let result = chunks[0];
-
-	for (let i = 1; i < chunks.length; i++) {
-		for (let j = result.length % spaceLen; j < spaceLen; j++) {
-			result += ' ';
-		}
-
-		result += chunks[i];
-	}
-	return result;
-};
-
-/**
- * @param a An indentation string
- * @param b A string of tabs and spaces to be compared with [a]
- * @return true iff [a] (an indentation string) is roughly equivalent to [b].
- */
-const isIndentationEquivalent = (state: EditorState, a: string, b: string): boolean => {
-	// Consider sublists to be the same as their parent list if they have the same
-	// label plus or minus 1 space.
-	return Math.abs(tabsToSpaces(state, a).length - tabsToSpaces(state, b).length) <= 1;
-};
-
-/**
- * Expands [sel] to the smallest container node with name in [nodeNames].
- * @param nodeNames Either a string containing the name of the node to expand to
- * 					in the syntax tree or a list of node names.
- * @return the expanded selection.
- * @see https://github.com/lezer-parser/markdown/blob/3c5f5dc3b5be08e19c32f0f7fac6f3619a58c911/src/markdown.ts#L41
- */
-const growSelectionToNode = (
-	state: EditorState, sel: SelectionRange, nodeNames: string|string[]
-): SelectionRange => {
-	let newFrom = null;
-	let newTo = null;
-	let smallestLen = Infinity;
-
-	const isAcceptableNode = (name: string): boolean => {
-		if (typeof nodeNames == 'string') {
-			return name == nodeNames;
-		}
-
-		for (const otherName of nodeNames) {
-			if (otherName == name) {
-				return true;
-			}
-		}
-
-		return false;
-	};
-
-	// Find the smallest range.
-	syntaxTree(state).iterate({
-		from: sel.from, to: sel.to,
-		enter: node => {
-			if (isAcceptableNode(node.name)) {
-				if (node.to - node.from < smallestLen) {
-					newFrom = node.from;
-					newTo = node.to;
-					smallestLen = newTo - newFrom;
-				}
-			}
-		},
-	});
-
-	// If it's in such a node,
-	if (newFrom != null && newTo != null) {
-		return EditorSelection.range(newFrom, newTo);
-	} else {
-		return sel;
-	}
-};
-
-/**
- * Toggles whether the given selection matches [spec].
- *
- * @param sel The region to add/remove formatting from.
- * @param spec Specifies how the target region starts/stops. To add formatting,
- * `spec.templateStart` is added before `sel` and `spec.templateStop` is added after.
- *
- * For example, something similar to surroundSelecton('**', '**') would surround
- * every selection range with asterisks (including the caret).
- * If the selection is already surrounded by these characters, they are
- * removed.
- *
- * @return A list of changes and an updated selection.
- */
-const toggleRegionSurrounded = (
-	doc: DocumentText, sel: SelectionRange, spec: RegionSpec
-): SelectionUpdate => {
-	let content = doc.sliceString(sel.from, sel.to);
-	const startMatchLen = spec.matchStart(doc, sel);
-	const endMatchLen = spec.matchStop(doc, sel);
-
-	const startsWithBefore = startMatchLen >= 0;
-	const endsWithAfter = endMatchLen >= 0;
-
-	const changes = [];
-	let finalSelStart = sel.from;
-	let finalSelStop = sel.to;
-
-	if (startsWithBefore && endsWithAfter) {
-		// Remove the before and after.
-		content = content.substring(startMatchLen);
-		content = content.substring(0, content.length - endMatchLen);
-
-		finalSelStop -= startMatchLen + endMatchLen;
-
-		changes.push({
-			from: sel.from,
-			to: sel.to,
-			insert: content,
-		});
-	} else {
-		changes.push({
-			from: sel.from,
-			insert: spec.templateStart,
-		});
-
-		changes.push({
-			from: sel.to,
-			insert: spec.templateStop,
-		});
-
-		// If not a caret,
-		if (!sel.empty) {
-			// Select the surrounding chars.
-			finalSelStop += spec.templateStart.length + spec.templateStop.length;
-		} else {
-			// Position the caret within the added content.
-			finalSelStart = sel.from + spec.templateStart.length;
-			finalSelStop = finalSelStart;
-		}
-	}
-
-	return {
-		changes,
-		range: EditorSelection.range(finalSelStart, finalSelStop),
-	};
-};
-
-/**
- * For all selections in the given `EditorState`, toggles whether each
- * is contained in a region of type [spec].
- * @param nodeName The name of the SyntaxNode associated with [spec].
- * @param spec Describes the formatting to toggle.
- * @return A TransactionSpec that toggles each selected region's formatting.
- * @see toggleSelectionFormat
- */
-const toggleGlobalSelectionFormat = (
-	state: EditorState, nodeName: string, spec: RegionSpec
-): TransactionSpec => {
-	const changes = state.changeByRange((sel: SelectionRange) => {
-		return toggleSelectionFormat(state, nodeName, sel, spec);
-	});
-	return changes;
-};
-
-const toggleSelectionFormat = (
-	state: EditorState, nodeName: string, sel: SelectionRange, spec: RegionSpec
-): SelectionUpdate => {
-	const endMatchLen = spec.matchStop(state.doc, sel);
-
-	// If at the end of the region, move the
-	// caret to the end.
-	// E.g.
-	//   **foobar|**
-	//   **foobar**|
-	if (sel.empty && endMatchLen > -1) {
-		const newCursorPos = sel.from + endMatchLen;
-
-		return {
-			range: EditorSelection.cursor(newCursorPos),
-		};
-	}
-
-	// Grow the selection to encompass the entire node.
-	const newRange = growSelectionToNode(state, sel, nodeName);
-	return toggleRegionSurrounded(state.doc, newRange, spec);
-};
-
-/**
- * Toggle formatting in a region, applying a block version of the formatting
- * if multiple lines are selected.
- *
- * @param inlineNodeName the `SyntaxNode::name` associated with the inline version
- * 	of the region.
- * @param inlineSpec specifies how to find/create the inline region.
- * @param blockNodeName the `SyntaxNode::name` associated with the **block** version
- * 	of the region (e.g. `FencedCode`).
- * @param blockTemplate The `start` and `stop` strings that can be used to surround the
- * 	selected lines to create a block region.
- * @param preserveBlockQuotes `true` iff block quote formatting (`> `...) should be ignored
- * 	when identifying region formatting.
- */
-export const toggleRegionFormat = (
-	state: EditorState,
-
-	inlineNodeName: string, inlineSpec: RegionSpec,
-
-	// The block version of the tag (e.g. fenced code)
-	blockNodeName: string,
-	blockRegex: { start: RegExp; stop: RegExp },
-
-	// start: Single-line content that precedes the block
-	// stop: Single-line content that follows the block.
-	// line breaks will be added
-	blockTemplate: { start: string; stop: string },
-
-	// If false, don't ensure that block formatting preserves block quotes
-	preserveBlockQuotes: boolean = true
-): TransactionSpec => {
-	const doc = state.doc;
-
-	const getMatchEndPoints = (
-		match: RegExpMatchArray, line: Line, inBlockQuote: boolean
-	): [startIdx: number, stopIdx: number] => {
-		const startIdx = line.from + match.index;
-		let stopIdx;
-
-		let contentLength = line.text.length;
-
-		// Don't treat '> ' as part of the line's content if we're in a blockquote.
-		if (inBlockQuote && preserveBlockQuotes) {
-			contentLength -= blockQuoteStartLen;
-		}
-
-		// If it matches the entire line, remove the newline character.
-		if (match[0].length == contentLength) {
-			stopIdx = line.to + 1;
-		} else {
-			stopIdx = startIdx + match[0].length;
-
-			// Take into account the extra '> ' characters, if necessary
-			if (inBlockQuote && preserveBlockQuotes) {
-				stopIdx += blockQuoteStartLen;
-			}
-		}
-
-		stopIdx = Math.min(stopIdx, doc.length);
-		return [startIdx, stopIdx];
-	};
-
-	// Returns a change spec that converts an inline region to a block region
-	// only if the user's cursor is in an empty inline region.
-	// For example,
-	//    $|$ -> $$\n|\n$$ where | represents the cursor.
-	const handleInlineToBlockConversion = (sel: SelectionRange) => {
-		if (!sel.empty) {
-			return null;
-		}
-
-		const startMatchLen = inlineSpec.matchStart(doc, sel);
-		const stopMatchLen = inlineSpec.matchStop(doc, sel);
-
-		if (startMatchLen >= 0 && stopMatchLen >= 0) {
-			const inlineStart = sel.from - startMatchLen;
-			const inlineStop = sel.from + stopMatchLen;
-
-			const fromLine = doc.lineAt(sel.from);
-			const inBlockQuote = fromLine.text.match(blockQuoteRegex);
-
-			let lineStartStr = '\n';
-			if (inBlockQuote && preserveBlockQuotes) {
-				lineStartStr = '\n> ';
-			}
-
-			// Determine the text that starts the new block (e.g. \n$$\n for
-			// a math block).
-			let blockStart = `${blockTemplate.start}${lineStartStr}`;
-			if (fromLine.from != inlineStart) {
-				// Add a line before to put the start of the block
-				// on its own line.
-				blockStart = lineStartStr + blockStart;
-			}
-
-			return {
-				changes: [
-					{
-						from: inlineStart,
-						to: inlineStop,
-						insert: `${blockStart}${lineStartStr}${blockTemplate.stop}`,
-					},
-				],
-
-				range: EditorSelection.cursor(inlineStart + blockStart.length),
-			};
-		}
-
-		return null;
-	};
-
-	const changes = state.changeByRange((sel: SelectionRange) => {
-		const blockConversion = handleInlineToBlockConversion(sel);
-		if (blockConversion) {
-			return blockConversion;
-		}
-
-		// If we're in the block version, grow the selection to cover the entire region.
-		sel = growSelectionToNode(state, sel, blockNodeName);
-
-		const fromLine = doc.lineAt(sel.from);
-		const toLine = doc.lineAt(sel.to);
-		let fromLineText = fromLine.text;
-		let toLineText = toLine.text;
-
-		let charsAdded = 0;
-		const changes = [];
-
-		// Single line: Inline toggle.
-		if (fromLine.number == toLine.number) {
-			return toggleSelectionFormat(state, inlineNodeName, sel, inlineSpec);
-		}
-
-		// Are all lines in a block quote?
-		let inBlockQuote = true;
-		for (let i = fromLine.number; i <= toLine.number; i++) {
-			const line = doc.line(i);
-
-			if (!line.text.match(blockQuoteRegex)) {
-				inBlockQuote = false;
-				break;
-			}
-		}
-
-		// Ignore block quote characters if in a block quote.
-		if (inBlockQuote && preserveBlockQuotes) {
-			fromLineText = fromLineText.substring(blockQuoteStartLen);
-			toLineText = toLineText.substring(blockQuoteStartLen);
-		}
-
-		// Otherwise, we're toggling the block version
-		const startMatch = blockRegex.start.exec(fromLineText);
-		const stopMatch = blockRegex.stop.exec(toLineText);
-		if (startMatch && stopMatch) {
-			// Get start and stop indicies for the starting and ending matches
-			const [fromMatchFrom, fromMatchTo] = getMatchEndPoints(startMatch, fromLine, inBlockQuote);
-			const [toMatchFrom, toMatchTo] = getMatchEndPoints(stopMatch, toLine, inBlockQuote);
-
-			// Delete content of the first line
-			changes.push({
-				from: fromMatchFrom,
-				to: fromMatchTo,
-			});
-			charsAdded -= fromMatchTo - fromMatchFrom;
-
-			// Delete content of the last line
-			changes.push({
-				from: toMatchFrom,
-				to: toMatchTo,
-			});
-			charsAdded -= toMatchTo - toMatchFrom;
-		} else {
-			let insertBefore, insertAfter;
-
-			if (inBlockQuote && preserveBlockQuotes) {
-				insertBefore = `> ${blockTemplate.start}\n`;
-				insertAfter = `\n> ${blockTemplate.stop}`;
-			} else {
-				insertBefore = `${blockTemplate.start}\n`;
-				insertAfter = `\n${blockTemplate.stop}`;
-			}
-
-			changes.push({
-				from: fromLine.from,
-				insert: insertBefore,
-			});
-
-			changes.push({
-				from: toLine.to,
-				insert: insertAfter,
-			});
-			charsAdded += insertBefore.length + insertAfter.length;
-		}
-
-		return {
-			changes,
-
-			// Selection should now encompass all lines that were changed.
-			range: EditorSelection.range(
-				fromLine.from, toLine.to + charsAdded
-			),
-		};
-	});
-
-	return changes;
-};
-
-/**
- * Toggles whether all lines in the user's selection start with [regex].
- * @param template is that match of [regex] that is used when adding a match.
- * @param matchEmpty if true, all lines **after the first** that have no non-space
- * content are ignored.
- * @param nodeName if given, is the name of the node to expand the selection
- * to, (e.g. TaskList to expand selections to containing TaskLists if possible).
- * Note that selection is only expanded if the existing selection is empty
- * (just a caret).
- */
-export const toggleSelectedLinesStartWith = (
-	state: EditorState,
-	regex: RegExp,
-	template: string,
-	matchEmpty: boolean, nodeName?: string, ignoreBlockQuotes: boolean = true
-): TransactionSpec => {
-	const getLineContentStart = (line: Line): number => {
-		if (!ignoreBlockQuotes) {
-			return line.from;
-		}
-
-		const blockQuoteMatch = line.text.match(blockQuoteRegex);
-		if (blockQuoteMatch) {
-			return line.from + blockQuoteMatch[0].length;
-		}
-
-		return line.from;
-	};
-
-	const getLineContent = (line: Line): string => {
-		const contentStart = getLineContentStart(line);
-		return line.text.substring(contentStart - line.from);
-	};
-
-	const changes = state.changeByRange((sel: SelectionRange) => {
-		// Attempt to select all lines in the region
-		if (nodeName && sel.empty) {
-			sel = growSelectionToNode(state, sel, nodeName);
-		}
-
-		const doc = state.doc;
-		const fromLine = doc.lineAt(sel.from);
-		const toLine = doc.lineAt(sel.to);
-		let hasProp = false;
-		let charsAdded = 0;
-
-		const changes = [];
-		const lines = [];
-
-		for (let i = fromLine.number; i <= toLine.number; i++) {
-			const line = doc.line(i);
-			const text = getLineContent(line);
-
-			// If already matching [regex],
-			if (text.search(regex) == 0) {
-				hasProp = true;
-			}
-
-			lines.push(line);
-		}
-
-		for (const line of lines) {
-			const text = getLineContent(line);
-			const contentFrom = getLineContentStart(line);
-
-			// Only process if the line is non-empty.
-			if (!matchEmpty && text.trim().length == 0
-				// Treat the first line differently
-				&& fromLine.number < line.number) {
-				continue;
-			}
-
-			if (hasProp) {
-				const match = text.match(regex);
-				if (!match) {
-					continue;
-				}
-				changes.push({
-					from: contentFrom,
-					to: contentFrom + match[0].length,
-					insert: '',
-				});
-
-				charsAdded -= match[0].length;
-			} else {
-				changes.push({
-					from: contentFrom,
-					insert: template,
-				});
-
-				charsAdded += template.length;
-			}
-		}
-
-		// If the selection is empty and a single line was changed, don't grow it.
-		// (user might be adding a list/header, in which case, selecting the just
-		// added text isn't helpful)
-		let newSel;
-		if (sel.empty && fromLine.number == toLine.number) {
-			const regionEnd = toLine.to + charsAdded;
-			newSel = EditorSelection.cursor(regionEnd);
-		} else {
-			newSel = EditorSelection.range(fromLine.from, toLine.to + charsAdded);
-		}
-
-		return {
-			changes,
-
-			// Selection should now encompass all lines that were changed.
-			range: newSel,
-		};
-	});
-
-	return changes;
-};
-
-/**
- * Ensures that ordered lists within [sel] are numbered in ascending order.
- * @return a list of changes and an updated selection after adjusting such lists.
- */
-export const renumberList = (state: EditorState, sel: SelectionRange): SelectionUpdate => {
-	const doc = state.doc;
-
-	const listItemRegex = /^(\s*)(\d+)\.\s?/;
-	const changes: ChangeSpec[] = [];
-	const fromLine = doc.lineAt(sel.from);
-	const toLine = doc.lineAt(sel.to);
-	let charsAdded = 0;
-
-	// Re-numbers ordered lists and sublists with numbers on each line in [linesToHandle]
-	const handleLines = (linesToHandle: Line[]) => {
-		let currentGroupIndentation = '';
-		let nextListNumber = 1;
-		const listNumberStack: number[] = [];
-		let prevLineNumber;
-
-		for (const line of linesToHandle) {
-			// Don't re-handle lines.
-			if (line.number == prevLineNumber) {
-				continue;
-			}
-			prevLineNumber = line.number;
-
-			const filteredText = stripBlockquote(line);
-			const match = filteredText.match(listItemRegex);
-			const indentation = match[1];
-
-			const indentationLen = tabsToSpaces(state, indentation).length;
-			const targetIndentLen = tabsToSpaces(state, currentGroupIndentation).length;
-			if (targetIndentLen < indentationLen) {
-				listNumberStack.push(nextListNumber);
-				nextListNumber = 1;
-			} else if (targetIndentLen > indentationLen) {
-				nextListNumber = listNumberStack.pop() ?? parseInt(match[2], 10);
-			}
-
-			if (targetIndentLen != indentationLen) {
-				currentGroupIndentation = indentation;
-			}
-
-			const from = line.to - filteredText.length;
-			const to = from + match[0].length;
-			const inserted = `${indentation}${nextListNumber}. `;
-			nextListNumber++;
-
-			changes.push({
-				from,
-				to,
-				insert: inserted,
-			});
-			charsAdded -= to - from;
-			charsAdded += inserted.length;
-		}
-	};
-
-	const linesToHandle: Line[] = [];
-	syntaxTree(state).iterate({
-		from: sel.from,
-		to: sel.to,
-		enter: (nodeRef: SyntaxNodeRef) => {
-			if (nodeRef.name == 'ListItem') {
-				for (const node of nodeRef.node.parent.getChildren('ListItem')) {
-					const line = doc.lineAt(node.from);
-					const filteredText = stripBlockquote(line);
-					const match = filteredText.match(listItemRegex);
-					if (match) {
-						linesToHandle.push(line);
-					}
-				}
-			}
-		},
-	});
-
-	linesToHandle.sort((a, b) => a.number - b.number);
-	handleLines(linesToHandle);
-
-	// Re-position the selection in a way that makes sense
-	if (sel.empty) {
-		sel = EditorSelection.cursor(toLine.to + charsAdded);
-	} else {
-		sel = EditorSelection.range(
-			fromLine.from,
-			toLine.to + charsAdded
-		);
-	}
-
-	return {
-		range: sel,
-		changes,
-	};
-};
 
 /** Bolds/unbolds the current selection. */
 export const toggleBolded: Command = (view: EditorView): boolean => {
 	logMessage('Toggling bolded!');
 
-	const changes = toggleGlobalSelectionFormat(view.state, 'StrongEmphasis', new RegionSpec({
-		templateStart: '**',
-		templateStop: '**',
-	}));
+	const spec = RegionSpec.of({ template: '**', nodeName: 'StrongEmphasis' });
+	const changes = toggleInlineFormatGlobally(view.state, spec);
 
 	view.dispatch(changes);
 	return true;
@@ -654,15 +34,12 @@ export const toggleBolded: Command = (view: EditorView): boolean => {
 export const toggleItalicized: Command = (view: EditorView): boolean => {
 	logMessage('Toggling italicized!');
 
-	const changes = toggleGlobalSelectionFormat(view.state, 'Emphasis', new RegionSpec({
-		// Template start/end
-		templateStart: '_',
-		templateStop: '_',
+	const changes = toggleInlineFormatGlobally(view.state, {
+		nodeName: 'Emphasis',
 
-		// Regular expressions that match all possible start/ends
-		startExp: /[_*]/g,
-		stopExp: /[_*]/g,
-	}));
+		template: { start: '_', end: '_' },
+		matcher: { start: /[_*]/g, end: /[_*]/g },
+	});
 	view.dispatch(changes);
 
 	return true;
@@ -677,19 +54,16 @@ export const toggleCode: Command = (view: EditorView): boolean => {
 	logMessage('Toggling code!');
 
 	const codeFenceRegex = /^```\w*\s*$/;
-	const inlineRegionSpec = new RegionSpec({
-		templateStart: '`',
-		templateStop: '`',
-	});
+	const inlineRegionSpec = RegionSpec.of({ template: '`', nodeName: 'InlineCode' });
+	const blockRegionSpec: RegionSpec = {
+		nodeName: 'FencedCode',
+		template: { start: '```', end: '```' },
+		matcher: { start: codeFenceRegex, end: codeFenceRegex },
+	};
 
-	const changes = toggleRegionFormat(
-		view.state,
-		'InlineCode', inlineRegionSpec,
-
-		'FencedCode', { start: codeFenceRegex, stop: codeFenceRegex },
-		{ start: '```', stop: '```' }
-	);
+	const changes = toggleRegionFormatGlobally(view.state, inlineRegionSpec, blockRegionSpec);
 	view.dispatch(changes);
+
 	return true;
 };
 
@@ -697,19 +71,18 @@ export const toggleMath: Command = (view: EditorView): boolean => {
 	logMessage('Toggling math!');
 
 	const blockStartRegex = /^\$\$/;
-	const blockStopRegex = /\$\$\s*$/;
-	const inlineRegionSpec = new RegionSpec({
-		templateStart: '$',
-		templateStop: '$',
+	const blockEndRegex = /\$\$\s*$/;
+	const inlineRegionSpec = RegionSpec.of({ nodeName: 'InlineMath', template: '$' });
+	const blockRegionSpec = RegionSpec.of({
+		nodeName: 'BlockMath',
+		template: '$$',
+		matcher: {
+			start: blockStartRegex,
+			end: blockEndRegex,
+		},
 	});
 
-	const changes = toggleRegionFormat(
-		view.state,
-		'InlineMath', inlineRegionSpec,
-
-		'BlockMath', { start: blockStartRegex, stop: blockStopRegex },
-		{ start: '$$', stop: '$$' }
-	);
+	const changes = toggleRegionFormatGlobally(view.state, inlineRegionSpec, blockRegionSpec);
 	view.dispatch(changes);
 
 	return true;
